Move inference dispatcher to infnode. Improve abstract atom comparison for inference testing result.
Partially fix test_while
